requirements.txt
accelerate==0.31.0
aioaws==0.13
aiobotocore==2.13.3
aiofiles==23.2.1
aiohttp==3.9.5
boto3==1.34.131
botocore==1.34.131
cachetools==5.5.2
celery==5.4.0
datasets==2.19.2
dill==0.3.8
dvc==3.51.2
dvc-data==3.15.2
dvc-s3==3.2.0
email-validator==2.1.2
evaluate==0.4.2
faiss-cpu==1.8.0
fastapi==0.111.0
flufl-lock==7.1.1
fsspec==2024.3.1
gunicorn==22.0.0
httpx==0.27.0
huggingface-hub==0.23.4
importlib-metadata==7.1.0
intel-openmp==2021.4.0
jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langgraph==0.0.69
langsmith==0.1.77
markdown==3.10
mkl==2021.4.0
mlflow==2.13.2
multiprocess==0.70.16
numpy==1.26.4
openai==1.109.1
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.5
packaging==24.2
pandas==2.2.2
platformdirs==3.11.0
prometheus-client==0.20.0
protobuf==4.25.8
pyarrow==15.0.2
pyarrow-hotfix==0.7
pydantic==2.7.4
pydantic-core==2.18.4
pysbd==0.3.4
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
pyyaml==6.0.1
querystring-parser==1.2.4
ragas==0.1.9
redis==5.0.6
regex==2024.5.15
requests==2.32.3
s3fs==2024.3.1
s3transfer==0.10.4
scikit-learn==1.5.0
sentence-transformers==3.0.1
starlette==0.37.2
tbb==2021.13.1
tenacity==8.4.1
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.0
ujson==5.11.0
uvicorn==0.30.1
peft==0.13.2
nltk==3.9.2
rouge==1.0.1